+++ src/main/java/pl/polskaamazonka/backend/repository/LinkCheckHistoryRepository.java
@@ -3,5 +3,5 @@ package pl.polskaamazonka.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.polskaamazonka.backend.model.Linkcheckhistory;
 
-public interface LinkCheckHistoryRepository extends JpaRepository<Linkcheckhistory, Integer> {
+public interface LinkCheckHistoryRepository extends JpaRepository<Linkcheckhistory, Long> {
 }
